undo r1713

+++ sources/session.php
@@ -190,9 +190,13 @@ class session {
 			//
 			// Remove outdated sessions and searches if needed
 			//
-			$min_updated = $current_time - ( $functions->get_config('session_max_lifetime') * 60 );
-			$add_to_remove_query[] = "updated < ".$min_updated;
-			$db->query("DELETE FROM ".TABLE_PREFIX."searches WHERE time < ".$min_updated);
+			if ( $functions->get_config('session_max_lifetime') ) {
+				
+				$min_updated = $current_time - ( $functions->get_config('session_max_lifetime') * 60 );
+				$add_to_remove_query[] = "updated < ".$min_updated;
+				$db->query("DELETE FROM ".TABLE_PREFIX."searches WHERE time < ".$min_updated);
+				
+			}
 			
 			//
 			// Remove sessions with banned IP addresses
